# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/add-friends": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "User Send Request",
                "description": "User Send Request",
                "operationId": "Users Send Request",
                "requestBody": {
                    "description": "user sent a friend pressure request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user sent a friend pressure request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/confirm-request": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "user accepts friend request",
                "description": "user accepts friend request",
                "operationId": "user accepts friend request",
                "requestBody": {
                    "description": "user accepts friend request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "properties": {
                                    "sender_id": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user accepts friend request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/cancel-request": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "user request canceled",
                "description": "user request canceled",
                "operationId": "user request canceled",
                "requestBody": {
                    "description": "user request canceled",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "properties": {
                                    "sender_id": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user request canceled",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/delete-friend": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "user removed from friends list",
                "description": "user removed from friends list",
                "operationId": "user removed from friends list",
                "requestBody": {
                    "description": "user removed from friends list",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ],
                                "properties": {
                                    "sender_id": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user removed from friends list",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/friends-birth": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "get users birthday",
                "description": "get users birthday",
                "operationId": "get user birthday",
                "requestBody": {
                    "description": "get users birthday",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "get users birthday",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/SinglePageholiday/id": {
            "post": {
                "tags": [
                    "SinglePageholiday"
                ],
                "summary": "SinglePageholiday",
                "description": "SinglePageholiday",
                "operationId": "SinglePageholiday",
                "requestBody": {
                    "description": "SinglePageholiday",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "true"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "get users birthday",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}